+++ app/src/main/kotlin/view/screens/MainScreenView.kt
@@ -8,11 +8,10 @@ import androidx.compose.foundation.gestures.scrollable
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import view.exceptionDialog.exceptionView
+import view.dialogs.exceptionView
 import view.graph.GraphView
 import viewmodel.colors.ColorTheme
 import viewmodel.screens.MainScreenViewModel
